refactor(types): create newtype for gas

+++ near-sdk/src/test_utils/context.rs
@@ -99,7 +99,7 @@ impl VMContextBuilder {
     }
 
     pub fn prepaid_gas(&mut self, gas: Gas) -> &mut Self {
-        self.context.prepaid_gas = gas;
+        self.context.prepaid_gas = u64::from(gas);
         self
     }
